Updated changes in robot code
